Bug Fixes
Please see the change_log.txt for details.
