--- ExCID_v2.0/Config.txt
@@ -1,2 +1,0 @@
-DataBaseDir=
-AnnotationDir=
