+++ ExCID_v2.1/Config.txt
@@ -1,0 +1,2 @@
+DataBaseDir=
+AnnotationDir=
